% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggGDPlifexp}
\alias{aggGDPlifexp}
\title{Aggregated on GDP, life expectancy, and population over time.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{continent}{Continent name}
  \item{lifeExp}{Average life expectancy at birth, in years}
  \item{pop}{Average population, in nb of people}
  \item{gdpPercap}{Average GDP per capita, in 2005-international dollars}
}
}
\source{
\url{https://github.com/jennybc/gapminder}
}
\usage{
aggGDPlifexp
}
\description{
The gapminder data on population size, life expectancy,
 and GDP per capita levels, aggregated over continents.
}
\keyword{datasets}
