% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beer}
\alias{beer}
\title{Data on beer consumption}
\format{
A data frame with 30 rows and 5 variables
\describe{
  \item{q}{litres of beer consumed}
  \item{pb}{Price of beer ($)}
  \item{pl}{price of other liquor ($)}
  \item{pr}{price of remaining goods and services (an index)}
  \item{i}{income ($)}
}
}
\source{
\url{http://www.principlesofeconometrics.com/poe4/poe4stata.htm}
}
\usage{
beer
}
\description{
A data set on beer consumption; the data was obtained for a single
 household
}
\keyword{datasets}
