% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataScienceExercises.R
\docType{package}
\name{DataScienceExercises}
\alias{DataScienceExercises}
\title{DataScienceExercises: data and exercises for a data science class}
\section{Description}{

This package contains example data sets and exercises used during the
 course "Data Science", taught in the spring semester 2022 at the
 University of Flensburg. Its main use is to provide the data sets used
 during the lecture and to provide exercises that are implemented using
 the `learnr` package. To access the exercises, please have a look at the
 respective tutorial on the course homepage
}

\section{Author and maintainer}{

Claudius Gräbner-Radkowitsch (\href{https://claudius-graebner.com/}{Website}
and \href{https://github.com/graebnerc}{Github})
}

\section{Course homepage}{

The most recent course homepage with all necessary information can be found
 at \url{https://datascience-euf-spring22.netlify.app/}
}

\section{Feedback}{

Please submit bug reports directly via the
 \href{https://github.com/graebnerc/DataScienceExercises/issues}{issue tracker}.
 Your help is greatly appreciated!
}

