% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUFstudents}
\alias{EUFstudents}
\title{Artificial data on heights of EUF students in 2021}
\format{
A data frame with 6402 rows and 2 variables
\describe{
  \item{Gender}{Gender of the person}
  \item{Height}{Height of the person}
}
}
\source{

}
\usage{
EUFstudents
}
\description{
A an artificial data set about the height of students from the Europa
 University Flensburg. The composition of the population was set according
 to official information about the student population in 2021, the heights
 were created artificially by drawing from a random distribution, which was
 calibrated for mean and standard deviation parameters of the German
 population.
}
\keyword{datasets}
