% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUFstudentsemesters}
\alias{EUFstudentsemesters}
\title{Study semester of the EUF students in 2021}
\format{
A data frame with 5875 rows and 1 variable
\describe{
  \item{ID}{An ID for the student}
  \item{Semester}{The study semester of the student}
}
}
\source{

}
\usage{
EUFstudentsemesters
}
\description{
A data set on the distribution of study semesters for all students of the
 Europa-University Flensburg in the spring semester 2021.
}
\keyword{datasets}
