% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdplifexp2007}
\alias{gdplifexp2007}
\title{Data on GDP, life expectancy, and population size in 2007.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{country}{Country name}
  \item{continent}{Continent name}
  \item{lifeExp}{Life expectancy at birth, in years}
  \item{pop}{Population in 2007, in nb of people}
  \item{gdpPercap}{GDP per capita, in 2005-international dollars}
}
}
\source{
\url{https://github.com/jennybc/gapminder}
}
\usage{
gdplifexp2007
}
\description{
A subset of the gapminder data on population size, life expectancy,
 and GDP per capita levels in 2007.
}
\keyword{datasets}
