% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{pop_var}
\alias{pop_var}
\title{Compute the population variance of a vector}
\usage{
pop_var(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
The population variance of this vector
}
\description{
Computes the population variance as defined by
\deqn{\sigma^2=\frac{1}{N}\sum_{i=1}^N(x_i-\mu)^2}
where \eqn{\mu} is the population mean (\eqn{\frac{1}{N}\sum_{i=1}^Nx_i})
and \eqn{N} is the size of the population
}
\examples{
pop_var(c(2, 3, 4, 5))
pop_var(rnorm(20, 1, 0))
}
