% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine2dine}
\alias{wine2dine}
\title{Data on wine properties}
\format{
A data frame with 6497 rows and 3 variables:
\describe{
  \item{alcohol}{The alcohol content of the wine}
  \item{quality}{An quality score between 0 and 10, based on sensory data}
  \item{residual sugar}{The amount of sugar left in the wine}
  \item{kind}{The kind of wine, i.e. red or white}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/wine-quality/}
}
\usage{
wine2dine
}
\description{
A data set containing information about the alcohol and residual sugar
 contained in white and red wines, as well as their quality
}
\keyword{datasets}
